Use director hostname instead of IP in the tile
Using the IP address means that if an Ops Manager user configures the hostname of their BOSH director the deployment will fail, as the zone is not set up correctly in BOSH's UAA.

The hostname accessor will fall back to the IP address if an Ops Manager user does not set a hostname for their BOSH director. This makes it safe in both cases.