Correct ports
Some of these ports are wrong; this table is referenced in many places on PivDocs.

ODB -> BOSH Director happens to talk to TCP 8443 (UAA) and 25555 (Director HTTPS)
BOSH Agent -> BOSH Director talks via NATS on TCP 4222.   The only thing talks 6868 these days is Ops Man to the BOSH Director when it does a bosh-init or bosh2 create-env to initialize the Director.